Update patch set 1
Patch Set 1:

Goodnight! I'm in need of some help, who, I'll be very grateful! Following;

I installed this Lineage OS 18.1 ROM in my Moto G7 power, the only problem that is happening is that it is not allowing localization in any app. I've researched it in some places, and it teaches you to choose an option "always allow" in each application, but this option doesn't exist. There is only one "allow while using app" option, but it doesn't work. Does anyone have this same problem?

Patch-set: 1
CC: Gerrit User 29372 <29372@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>